requirements.txt
# API and Web
flask==2.3.3
werkzeug==2.3.7
gunicorn==20.1.0

# Data and ML
numpy
openai==1.3.7
qdrant-client==1.6.0

# Notion integration
notion-client==2.0.0

# Utilities
python-dotenv==1.0.0
requests==2.28.2
# uuid==1.30  # Removed as Python has UUID built in
pydantic>=1.10.8,<3  # This satisfies both openai and qdrant-client

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0

# Monitoring
prometheus-client==0.16.0

# Add this line or update existing httpx entry
httpx>=0.24.0